Feat(Lotto) : 로또의 번호가 중복되거나 로또의 길이가 6이 아니라면 예외 처리한다
