+++ docs/README.md
@@ -27,6 +27,7 @@
 
 ### 예외처리
 - [x] 구입 금액이 1,000원으로 나누어 떨어지지 않는 경우 예외 처리한다.
-- [ ] 생성된 로또의 번호가 중복되면 예외 처리한다.
+- [x] 생성된 로또의 번호가 중복되면 예외 처리한다.
+- [x] 생성된 로또의 길이가 6이 아니라면 예외 처리한다.
 - [ ] 당첨 로또에 보너스 번호가 포함되면 예외 처리한다.
 - [x] 로또 번호의 숫자 범위가 1 ~ 45 사이를 벗어나면 예외 처리 한다.
